Add configuration files for alacritty, skhdrc and starship.
I decided to try yabai and move from kitty/iterm to alacritty. The
performance difference is huge at least compared to iterm.

Yabai is used for window management, which is essential. This might set
me into to a path to not use tmux as much.

SKHD is a hotkey daemon. Very simple and works well with yabai (from the
same author).

Remove fish shell greet message and add padding to tmux status bar.