nvim: set buf formatting to always async
Previously when typescript lang server was loaded the buf formatting was
done syncronously, which resulted in poor dev experience due to not be
able to move around or run commands. This is a problem with bigger
repositories. Another solution is to use prettier_d, or using prettier_d
is not ruled out with this either. (prettier_d is preferred for faster
speed)