nvim: null-ls only attach when needed
Add attach_to table with appropriate should_attach functions.
Key is represented as the filetype and the value is what attach function
should be ran. The function should return true or false.

Add sh with function that always attaches null-ls.

Add default that will be ran if filetype is not found in the attach_to
map. By default we check for package.json and attach if package.json is
found. This way we do not cause a conflict between denols and tsserver.