nvim: switch `null-ls` to `efm`
`null-ls` was archived
https://github.com/jose-elias-alvarez/null-ls.nvim/issues/1621

This is a response to the archive decision as I will need a reliable
tool for formatting and linting that is going to be maintained in the
nearby future. Before switching to `null-ls` I was using `diagnosticls`,
which was quite slow and not fun due to needing to maintain hand made
configurations. `null-ls` on the other hand provided easy access to all
the linter and formatting configuration I needed. So `diagnosticls` was
not on the table when I was deciding what tool to use to replace
`null-ls`. Recently I stumbled upon `efm` and `efmls-configs` that
seemed to be exactly what I need.

`efm` is a general purpose language server written in go. At first
glance it seems to be a lot faster than the `null-ls` variant that was
written in lua. For example previously I was experiencing severe
slowdowns with eslint and multiple nodejs instances that were spun up,
at that time I didn't really connect the dots that `null-ls` was causing
the slowdows as the slowdowns seem not to be the case with `efm`. Even
my rust setup feels a lot faster now..?

`efmls-configs` provide "builtins" for all the common formatters and
linters, which makes maintaining and setting up the editor more _fun_.
One caveat here compared to `null-ls` is that there's no great way to
extend the command arguments, and if other arguments are needed a custom
config object has to be created.