Add initial `.aerospace.toml` config
The configuration is divided into three modes:

1. The main mode, which essentially is the entry point for other modes
2. The focus mode, which is used for focusing, moving and closing windows
3. The workspace mode, which is used for changing workspaces

The focus and workspace modes are custom modes. Both custom modes can be
exited by pressing 'q'.

When changing modes a file in the home directory called
`.aerospace.mode` is updated with the corresponding mode. This is
stateful information that can be utilized with integrations.