Improve aerospace window squeeze toggle
Previously the left and right padding was fixed to 600px, but this
created an unusable small window on smaller screens. Adjusted the script
to collect screen information using osascript. The padding is now
calculated from the screen width by dividing the whole screen width with
5. Essentially the window will take a space of 3/5 in the center of the
screen regardless of the actual screen size.