Add tomb cli and tomb sketchybar plugin
Tomb CLI https://github.com/erikjuhani/tomb is a personal task manager
that helps with task management that is using a single markdown file.

This is a personal workflow that I have found helpful and fun to do.

The sketchybar plugin showcases how many tasks I have for that
particular day and reminds me if I have not set any tasks for that day,
which usually means that I have forgotten to update my tasks.