Workflows: Use personal token to trigger workflows
In order to trigger the push event for the pull-request opened we have
to use a personal token (or some other token) to allow other workflows
to be triggered by an action.

Also the permissions should be elevated to allow writing of contents and
pull-requests. The default tokens were changed to read-only on
repositories created after Feb 2023. However this repository was created
prior.

> Pull requests created by the action using the default GITHUB_TOKEN
> cannot trigger other workflows. If you have on: pull_request or on: push
> workflows acting as checks on pull requests, they will not run.
> https://github.com/peter-evans/create-pull-request/blob/main/docs/concepts-guidelines.md#triggering-further-workflow-runs