Aerospace: Simplify configuration to one "main" binding layer
Previously I had 3 binding layers, main, focus and workspace. However,
in action this was suboptimal as I had a constant pause of trying to
understand which mode was on and trying to remember what the keys were.
This resulted me in _not_ using most of the commands and caused clunky
navigation between apps and windows.

I was thinking this could work like vim insert, visual modes, but
clearly it was not the case.

I simplified all the configuration under one binding layer "main" (the
default one) and the tiling management already feels much better.

I mainly use the hyper key `cmd-ctrl-alt-shift` to manage the windows.
However some workspace move commands are found under `alt-shift` as
these are more explicit operations and not done as frequently.