Show all workspaces in sketchybar with details of open windows
This required changing the aerospace configuration to send
workspace_change events when workspace or window focus is changed.

Changing focus also means opening or closing the windows.

The aerospace.sh plugin was modifed to show tabs for each workspace.
Initially the workspace number is shown. If one window is open in the
workspace the windows' app icon and name is displayed. If there are
multiple windows open in the single workspace each windows' app icon is
displayed in the tab.

Due to this change the currently opened window plugin was commented out
since the information is now contained within the tab system.

The currently focused workspace is shown with magenta highlight.